Move BasicBtnStyle struct to a separate model
